pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Cross-Server-Chat"
version = "0.1.0"
description = ""
authors = ["mel0n7 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
discord = "^1.7.3"
Flask = "^2.0.3"
mysql-connector-python = "^8.0.28"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
